The Muxer
This code is a GUI application built with PySide2 (a set of Python bindings for the Qt framework) that can mux (merge) video and audio files into a single file using the FFmpeg command-line tool.

The code defines a class called "Muxer" which inherits from the QWidget class and creates a graphical user interface. The GUI consists of buttons for selecting a video file, an audio file, and for running the muxing process. The user can select the files by clicking the "Upload Video" and "Upload Audio" buttons which open a file dialog. After the files are selected, the user can run the muxing process by clicking the "Run" button which opens another file dialog for naming the output file. If the video and audio files are both selected, the application runs the FFmpeg command using Python's subprocess.run function and displays a message box indicating whether the muxing process was successful or not.